error handling + bugfixes
