+++ src/client/pages/Checkout/CheckoutPayment/index.tsx
@@ -1,15 +1,19 @@
 import React from 'react'
 import { useQuoteCartData } from 'utils/hooks/useQuoteCartData'
 import { LoadingPage } from 'components/LoadingPage'
-import { CheckoutErrorModal } from '../shared/ErrorModal'
+import { CheckoutErrorModal, onRetry } from '../shared/ErrorModal'
 import { CheckoutPayment } from './CheckoutPayment'
 
 export const Checkout = () => {
   const data = useQuoteCartData()
+
+  if (data?.error) {
+    console.error('Quote cart data error: no data')
+    return <CheckoutErrorModal isVisible onRetry={onRetry} />
+  }
   if (!data || data.loading) {
     return <LoadingPage loading />
   }
-
   const {
     bundleVariants,
     priceData,
@@ -18,21 +22,7 @@ export const Checkout = () => {
     quoteIds,
     selectedQuoteBundleVariant,
     checkoutStatus,
-    error,
   } = data
-  if (!data || data.loading) {
-    return <LoadingPage loading />
-  }
-
-  const onRetry = () => {
-    window.location.reload()
-    return false
-  }
-
-  if (error) {
-    console.error('Quote cart data error:', error.message, error)
-    return <CheckoutErrorModal isVisible onRetry={onRetry} />
-  }
 
   return (
     <CheckoutPayment
